+++ graphics/src/main/java/pl/warp/engine/graphics/pipeline/PipelineElement.java
@@ -5,7 +5,7 @@ package pl.warp.engine.graphics.pipeline;
  *         Created 2016-06-26 at 13
  */
 public interface PipelineElement {
-    void update(long delta);
+    void update(int delta);
     void init();
     void destroy();
     void onResize(int newWidth, int newHeight);
